requirements.txt
Flask
openai==0.28.0
python-dotenv==1.0.1
google-cloud-secret-manager
gunicorn
uvicorn
